# api/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Rin HTTP API",
    "version": "0.7.0",
    "description": "Authoritative preview wire contract for Rin Protocol v2. Bearer authentication is required only when the server is configured with a token; liveness and readiness probes are always unauthenticated."
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:7374"
    }
  ],
  "paths": {
...
    "/v2/diagnostics": {
      "get": {
        "operationId": "diagnostics",
        "x-rin-sdk-profile": "operational",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Diagnostics"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/v2/session/observe": {
      "post": {
        "operationId": "observe",
        "requestBody": {
          "$ref": "#/components/requestBodies/Observe"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MutationResult"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "507": {
            "$ref": "#/components/responses/InsufficientStorage"
          }
        }
      }
    },
    "/v2/agent/propose": {
      "post": {
        "operationId": "propose",
        "requestBody": {
          "$ref": "#/components/requestBodies/Propose"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ProposalResult"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "408": {
            "$ref": "#/components/responses/RequestTimeout"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "507": {
            "$ref": "#/components/responses/InsufficientStorage"
          }
        }
      }
    },
...
    "/v2/world/arbitrate": {
      "post": {
        "operationId": "arbitrate",
        "requestBody": {
          "$ref": "#/components/requestBodies/Arbitrate"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ArbitrationResult"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "507": {
            "$ref": "#/components/responses/InsufficientStorage"
          }
        }
      }
    },
...
    "/v2/session/snapshot": {
      "post": {
        "operationId": "snapshot",
        "requestBody": {
          "$ref": "#/components/requestBodies/Session"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Snapshot"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "507": {
            "$ref": "#/components/responses/InsufficientStorage"
          }
        }
      }
    },
    "/v2/session/restore": {
      "post": {
        "operationId": "restore",
        "requestBody": {
          "$ref": "#/components/requestBodies/Restore"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MutationResult"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "507": {
            "$ref": "#/components/responses/InsufficientStorage"
          }
        }
      }
    },
...
    "/v2/session/timeline": {
      "post": {
        "operationId": "timeline",
        "requestBody": {
          "$ref": "#/components/requestBodies/Timeline"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Timeline"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Used when the Rin server is configured with an authentication token."
      }
    },
...
    "requestBodies": {
...
      "Observe": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ObserveRequest"
            }
          }
        }
      },
      "Propose": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProposeRequest"
            }
          }
        }
      },
...
      "Arbitrate": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArbitrateRequest"
            }
          }
        }
      },
      "Session": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SessionRequest"
            }
          }
        }
      },
...
      "Restore": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RestoreRequest"
            }
          }
        }
      },
      "Timeline": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TimelineRequest"
            }
          }
        }
      },
...
    },
    "responses": {
...
      "Diagnostics": {
        "description": "Bounded content-free operational diagnostics.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DiagnosticsResponse"
            }
          }
        }
      },
...
      "MutationResult": {
        "description": "Mutation result.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MutationResultResponse"
            }
          }
        }
      },
      "ProposalResult": {
        "description": "Game-authorized proposal.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProposalResultResponse"
            }
          }
        }
      },
...
      "ArbitrationResult": {
        "description": "Arbitration result.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArbitrationResultResponse"
            }
          }
        }
      },
...
      "Snapshot": {
        "description": "Portable session snapshot.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SnapshotResponse"
            }
          }
        }
      },
...
      "Timeline": {
        "description": "Redacted timeline page.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TimelineResponseEnvelope"
            }
          }
        }
      },
...
      "BadRequest": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "Unauthorized": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "NotFound": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "RequestTimeout": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "Conflict": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "RequestTooLarge": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "UnsupportedMediaType": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "UnprocessableEntity": {
        "$ref": "#/components/responses/ErrorResponse"
      },
...
      "InsufficientStorage": {
        "$ref": "#/components/responses/ErrorResponse"
      },
      "InternalError": {
        "$ref": "#/components/responses/ErrorResponse"
      },
...
      "ErrorResponse": {
        "description": "Bounded protocol error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "DiagnosticsResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "properties": {
              "data": {
                "$ref": "#/components/schemas/DiagnosticsData"
              }
            }
          }
        ]
      },
      "MutationResultResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "properties": {
              "data": {
                "$ref": "#/components/schemas/MutationResult"
              }
            }
          }
        ]
      },
...
    }
  },
  "security": [
    {},
    {
      "bearerAuth": []
    }
  ]
}